:white_check_mark: test: Merge extended js families test to base version
